refactor: load cards list in selector function
